Add files via upload
The script was created using the eclipse.exe editor, please include all jar files needed to load the project. 

The Project May25 has to be renamed to any project name you choose to add this .java file and run .

please place the chromedriver.exe in the same directory as the .java file, and change its path in the .java file.